Rename the dakota_base module base
